# docs/swagger.yaml
basePath: /
definitions:
  domain.AnomalyCheckRequest:
    properties:
      durationMs:
        example: 5
        type: integer
      endpoint:
        example: AiPromptSyncScheduler.syncAiPromptsToDify
        type: string
      service:
        example: twdiw-customer-service-prod
        type: string
      timestampNano:
        example: 1737000000000000000
        type: integer
    type: object
  domain.AnomalyCheckResponse:
    properties:
      baseline:
        $ref: '#/definitions/domain.BaselineStats'
      baselineSource:
        allOf:
        - $ref: '#/definitions/domain.BaselineSource'
        example: exact
      bucket:
        $ref: '#/definitions/domain.TimeBucket'
      cannotDetermine:
        example: false
        type: boolean
      explanation:
        example: duration 5ms within threshold 2.00ms
        type: string
      fallbackLevel:
        example: 1
        type: integer
      isAnomaly:
        example: false
        type: boolean
      sourceDetails:
        example: 'exact match: 9|weekday'
        type: string
    type: object
  domain.AvailableServicesResponse:
    properties:
      services:
        items:
          $ref: '#/definitions/domain.ServiceEndpoint'
        type: array
      totalEndpoints:
        example: 17
        type: integer
      totalServices:
        example: 4
        type: integer
    type: object
  domain.BaselineSource:
    enum:
    - exact
    - nearby
    - daytype
    - global
    - unavailable
    type: string
    x-enum-comments:
      SourceDayType: 'Level 3: All hours of same day type'
      SourceExact: 'Level 1: Exact hour|dayType match'
      SourceGlobal: 'Level 4: All data (any hour, any day type)'
      SourceNearby: 'Level 2: Nearby hours (±1, ±2)'
      SourceUnavailable: 'Level 5: No data available'
    x-enum-descriptions:
    - 'Level 1: Exact hour|dayType match'
    - 'Level 2: Nearby hours (±1, ±2)'
    - 'Level 3: All hours of same day type'
    - 'Level 4: All data (any hour, any day type)'
    - 'Level 5: No data available'
    x-enum-varnames:
    - SourceExact
    - SourceNearby
    - SourceDayType
    - SourceGlobal
    - SourceUnavailable
  domain.BaselineStats:
    properties:
      mad:
        example: 0
        type: number
      p50:
        example: 1
        type: number
      p95:
        example: 2
        type: number
      sampleCount:
        example: 188
        type: integer
      updatedAt:
        example: "2026-01-16T02:00:00Z"
        type: string
    type: object
  domain.ErrorDetail:
    properties:
      code:
        example: trace_not_found
        type: string
      details:
        additionalProperties: true
        type: object
      message:
        example: Trace not found in Tempo
        type: string
    type: object
  domain.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/domain.ErrorDetail'
    type: object
  domain.LongestSpanResponse:
    properties:
      computedAt:
        example: "2026-01-20T10:12:35.001Z"
        type: string
      longestSpan:
        $ref: '#/definitions/domain.SpanSummary'
      source:
        example: tempo
        type: string
      traceID:
        example: abc123def456
        type: string
    type: object
  domain.ServiceEndpoint:
    properties:
      buckets:
        example:
        - 6|weekday
        - 9|weekday
        - 10|weekday
        - 12|weekday
        - 13|weekend
        - 17|weekday
        - 20|weekday
        items:
          type: string
        type: array
      endpoint:
        example: AiPromptSyncScheduler.syncAiPromptsToDify
        type: string
      service:
        example: twdiw-customer-service-prod
        type: string
    type: object
  domain.SpanSummary:
    properties:
      durationMs:
        example: 842
        type: integer
      endTime:
        example: "2026-01-20T10:12:33.965Z"
        type: string
      name:
        example: db.query
        type: string
      parentSpanId:
        example: a1b2c3d4e5f6a7b8
        type: string
      service:
        example: orders-db
        type: string
      spanId:
        example: b7ad6b7169203331
        type: string
      startTime:
        example: "2026-01-20T10:12:33.123Z"
        type: string
    type: object
  domain.TimeBucket:
    properties:
      dayType:
        description: '"weekday" or "weekend"'
        example: weekday
        type: string
      hour:
        description: 0-23
        example: 9
        type: integer
    type: object
  domain.TraceEvent:
    properties:
      durationMs:
        example: 250
        type: integer
      rootServiceName:
        example: my-service
        type: string
      rootTraceName:
        example: GET /api/users
        type: string
      startTimeUnixNano:
        example: "1673000000000000000"
        type: string
      traceID:
        example: abc123def456
        type: string
    type: object
  domain.TraceLookupResponse:
    properties:
      count:
        example: 2
        type: integer
      end:
        example: 1736931600
        type: integer
      endpoint:
        example: GET /api/users
        type: string
      service:
        example: api-gateway
        type: string
      start:
        example: 1736928000
        type: integer
      traces:
        items:
          $ref: '#/definitions/domain.TraceEvent'
        type: array
    type: object
  handlers.HealthResponse:
    properties:
      status:
        example: ok
        type: string
    type: object
  store.Baseline:
    properties:
      mad:
        example: 43
        type: number
      p50:
        example: 233.5
        type: number
      p95:
        example: 562
        type: number
      sampleCount:
        example: 50
        type: integer
      updatedAt:
        example: "2026-01-15T08:00:00Z"
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: |-
    Time-Aware API Latency Anomaly Detection Service based on Grafana Tempo traces

    This service provides real-time latency anomaly detection using statistical methods (P50, P95, MAD).
    It automatically ingests traces from Tempo and maintains time-bucketed baselines per (service, endpoint, hour, dayType).
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Tempo Latency Anomaly Detection Service API
  version: "1.0"
paths:
  /healthz:
    get:
      consumes:
      - application/json
      description: Check if the service is running and healthy
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.HealthResponse'
      summary: Health check
      tags:
      - Health
  /v1/anomaly/check:
    post:
      consumes:
      - application/json
      description: |-
        Evaluate if a given request latency is anomalous based on historical baselines
        The service uses time-bucketed baselines (per hour and day type) and statistical methods (P50, P95, MAD) to detect anomalies
        Fallback: Automatically applies a 5-level strategy to select a baseline
        (exact → nearby → daytype → global → unavailable) when the exact bucket has insufficient data.
        Response fields:
        - baselineSource: which source was used (exact/nearby/daytype/global/unavailable)
        - fallbackLevel: numeric level 1-5 mapping to the source (1=exact, 2=nearby, 3=daytype, 4=global, 5=unavailable)
        - sourceDetails: human-readable details about the selected source (e.g. "exact match: 17|weekday")
        - cannotDetermine: true if no sufficient baseline exists to decide
      parameters:
      - description: Anomaly check request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.AnomalyCheckRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Includes baselineSource, fallbackLevel, sourceDetails, cannotDetermine
          schema:
            $ref: '#/definitions/domain.AnomalyCheckResponse'
        "400":
          description: Invalid JSON
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service not available
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Check for latency anomaly
      tags:
      - Anomaly Detection
  /v1/available:
    get:
      consumes:
      - application/json
      description: |-
        Retrieve all services and endpoints that have sufficient baseline samples for anomaly detection
        Returns a list of service/endpoint pairs grouped by service, along with their available time buckets
        Only includes baselines with sufficient samples (configured via min_samples)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.AvailableServicesResponse'
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service not available
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List available services and endpoints
      tags:
      - Available Services
  /v1/baseline:
    get:
      consumes:
      - application/json
      description: |-
        Retrieve baseline statistics (P50, P95, MAD) for a specific service, endpoint, hour, and day type
        Used for debugging and monitoring baseline data
      parameters:
      - description: Service name
        example: '"twdiw-customer-service-prod"'
        in: query
        name: service
        required: true
        type: string
      - description: Endpoint name
        example: '"GET /actuator/health"'
        in: query
        name: endpoint
        required: true
        type: string
      - description: Hour of day (0-23)
        example: 16
        in: query
        name: hour
        required: true
        type: integer
      - description: Day type (weekday or weekend)
        example: '"weekday"'
        in: query
        name: dayType
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Baseline'
        "400":
          description: Invalid parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Baseline not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Store not available
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get baseline statistics
      tags:
      - Baseline
  /v1/traces:
    get:
      consumes:
      - application/json
      description: Retrieve trace IDs and metadata for a service/endpoint within a
        time range
      parameters:
      - description: Service name
        example: '"api-gateway"'
        in: query
        name: service
        required: true
        type: string
      - description: Endpoint name
        example: '"GET /api/users"'
        in: query
        name: endpoint
        required: true
        type: string
      - description: Start time (unix seconds)
        example: 1736928000
        in: query
        name: start
        required: true
        type: integer
      - description: End time (unix seconds)
        example: 1736931600
        in: query
        name: end
        required: true
        type: integer
      - description: Max traces to return
        example: 200
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.TraceLookupResponse'
        "400":
          description: Invalid parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Tempo error
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Tempo not available
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Look up traces
      tags:
      - Traces
  /v1/traces/{traceId}/longest-span:
    get:
      consumes:
      - application/json
      description: Retrieve the longest span detail for a given trace ID
      parameters:
      - description: Trace ID
        example: "abc123def456"
        in: path
        name: traceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.LongestSpanResponse'
        "400":
          description: Invalid trace ID
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Trace not found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "422":
          description: Trace has no spans
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "502":
          description: Tempo error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "504":
          description: Tempo timeout
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "503":
          description: Tempo not available
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Get longest span in a trace
      tags:
      - Traces
schemes:
- http
- https
swagger: "2.0"
tags:
- description: Health check endpoints
  name: Health
- description: Anomaly detection and evaluation endpoints
  name: Anomaly Detection
- description: Baseline statistics query endpoints
  name: Baseline
- description: Query available services and endpoints with sufficient baseline data
  name: Available Services
- description: Trace lookup endpoints
  name: Traces